fix(bridge): do not display chain selector when source chain not supported
