+++ apps/cowswap-frontend/src/modules/tokensList/hooks/useChainsToSelect.ts
@@ -1,8 +1,8 @@
 import { useMemo } from 'react'
 
-import { BaseChainInfo, CHAIN_INFO, RPC_URLS } from '@cowprotocol/common-const'
+import { CHAIN_INFO } from '@cowprotocol/common-const'
 import { useIsBridgingEnabled } from '@cowprotocol/common-hooks'
-import { SupportedChainId, ChainInfo } from '@cowprotocol/cow-sdk'
+import { ChainInfo, SupportedChainId } from '@cowprotocol/cow-sdk'
 import { useIsSmartContractWallet, useWalletInfo } from '@cowprotocol/wallet'
 
 import { useBridgeSupportedNetworks } from 'entities/bridgeProvider'
@@ -12,43 +12,7 @@ import { Field } from 'legacy/state/types'
 import { useSelectTokenWidgetState } from './useSelectTokenWidgetState'
 
 import { ChainsToSelectState } from '../types'
-
-export function mapChainInfo(chainId: SupportedChainId, info: BaseChainInfo): ChainInfo {
-  return {
-    addressPrefix: info.addressPrefix,
-    contracts: {},
-    docs: {
-      url: info.docs,
-      name: `${info.label} Docs`,
-    },
-    isTestnet: chainId === SupportedChainId.SEPOLIA,
-    rpcUrls: {
-      default: {
-        http: [RPC_URLS[chainId]],
-      },
-    },
-    website: {
-      url: info.infoLink,
-      name: info.label,
-    },
-    id: chainId,
-    label: info.label,
-    nativeCurrency: {
-      ...info.nativeCurrency,
-      name: info.nativeCurrency.name || '',
-      symbol: info.nativeCurrency.symbol || '',
-    },
-    blockExplorer: {
-      url: info.explorer,
-      name: info.explorerTitle,
-    },
-    logo: {
-      light: info.logo.light,
-      dark: info.logo.dark,
-    },
-    color: info.color,
-  }
-}
+import { mapChainInfo } from '../utils/mapChainInfo'
 
 const SUPPORTED_CHAINS: ChainInfo[] = Object.keys(CHAIN_INFO).map((chainId) => {
   const supportedChainId = +chainId as SupportedChainId
@@ -74,15 +38,35 @@ export function useChainsToSelect(): ChainsToSelectState | undefined {
     if (!field || !isBridgingEnabled) return undefined
 
     const currentChainInfo = mapChainInfo(chainId, CHAIN_INFO[chainId])
-    const isCurrentChainInBridge = bridgeSupportedNetworks?.find((bridgeChain) => bridgeChain.id === chainId)
-    const chains =
-      field === Field.INPUT
-        ? SUPPORTED_CHAINS
-        : [...(isCurrentChainInBridge ? [] : [currentChainInfo]), ...(bridgeSupportedNetworks || [])]
+    const isSourceChainSupportedByBridge = Boolean(
+      bridgeSupportedNetworks?.find((bridgeChain) => bridgeChain.id === chainId),
+    )
+
+    // For the sell token selector we only display supported chains
+    if (field === Field.INPUT) {
+      return {
+        defaultChainId: selectedTargetChainId,
+        chains: SUPPORTED_CHAINS,
+        isLoading: false,
+      }
+    }
+
+    /**
+     * When the source chain is not supported by bridge provider
+     * We act as non-bridge mode
+     */
+    if (!isSourceChainSupportedByBridge) {
+      return {
+        defaultChainId: selectedTargetChainId,
+        chains: [],
+        isLoading: false,
+      }
+    }
 
     return {
       defaultChainId: selectedTargetChainId,
-      chains,
+      // Add the source network to the list if it's not supported by bridge provider
+      chains: [...(isSourceChainSupportedByBridge ? [] : [currentChainInfo]), ...(bridgeSupportedNetworks || [])],
       isLoading,
     }
   }, [field, selectedTargetChainId, chainId, bridgeSupportedNetworks, isLoading, isBridgingEnabled])
